requirements.txt
einops==0.8.1
ipython==8.12.3
librosa==0.10.1
numpy==2.3.5
pretty_midi==0.2.11
pyfluidsynth==1.3.4
python-dotenv==1.2.1
PyYAML==6.0.1
PyYAML==6.0.3
scipy==1.16.3
sf2utils==1.0.0
soundfile==0.12.1
torch==2.9.1
torchaudio==2.9.1
torchcrepe==0.0.18
torchvision==0.24.1
tqdm==4.66.1
transformers==4.48.0
